refactor: 플레이어 hit or stand 입력에 따른 로직 변경을 위해 컨트롤러 생성
